+++ NEWS.md
@@ -1,3 +1,7 @@
 # ggoutlines 0.0.0.9000
 
+* New function: `geom_line_outline()`.
+* New function: `geom_path_outline()`.
+* New function: `element_line_outline()`.
+* New function: `grob_outline_line()`.
 * Added a `NEWS.md` file to track changes to the package.